backend/requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
selenium==4.15.2
webdriver-manager==4.0.1
sentence-transformers>=2.7.0
huggingface_hub>=0.20.0
chromadb>=0.5.0
numpy<2.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
openai>=1.12.0
Pillow==10.1.0
